Back / 상품 엔티티 수정

+++ backend/src/main/java/com/example/backend/dto/product/BuyingDto.java
@@ -3,6 +3,7 @@ package com.example.backend.dto.product;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -14,5 +15,5 @@ public class BuyingDto {
     private Long buyingId;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime buyingBiddingTime;
-    private Long buyingBiddingPrice;
+    private BigDecimal buyingBiddingPrice;
 }
