+++ backend/src/main/java/com/example/backend/dto/product/Detail/BuyingHopeDto.java
@@ -2,6 +2,8 @@ package com.example.backend.dto.product.Detail;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @Setter
@@ -9,6 +11,6 @@ import lombok.*;
 @NoArgsConstructor
 public class BuyingHopeDto {
     private String productSize;
-    private Long buyingBiddingPrice;
+    private BigDecimal buyingBiddingPrice;
     private int buyingQuantity;
 }
