+++ backend/src/main/java/com/example/backend/dto/product/Detail/ProductsContractListDto.java
@@ -3,6 +3,7 @@ package com.example.backend.dto.product.Detail;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class ProductsContractListDto {
     private String productSize;
-    private Long productContractPrice;
+    private BigDecimal productContractPrice;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime productContractDate;
 }
