+++ backend/src/main/java/com/example/backend/dto/product/Detail/RecentlyPriceDto.java
@@ -3,6 +3,7 @@ package com.example.backend.dto.product.Detail;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -11,13 +12,13 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RecentlyPriceDto {
-    private Long latestPrice;
+    private BigDecimal latestPrice;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime latestDate;
-    private Long previousPrice;
+    private BigDecimal previousPrice;
     private Double changePercentage;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime salesBiddingTime;
-    private Long salesBiddingPrice;
+    private BigDecimal salesBiddingPrice;
     private Long differenceContract;
 }
