+++ backend/src/main/java/com/example/backend/dto/product/Detail/SalesBiddingDto.java
@@ -3,6 +3,7 @@ package com.example.backend.dto.product.Detail;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -15,10 +16,10 @@ public class SalesBiddingDto {
     private Long productId;
     private String modelNum;
     private String productSize;
-    private Long latestPrice;
-    private Long previousPrice;
+    private BigDecimal latestPrice;
+    private BigDecimal previousPrice;
     private Double previousPercentage;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
     private LocalDateTime salesBiddingTime;
-    private Long salesBiddingPrice;
+    private BigDecimal salesBiddingPrice;
 }
