+++ backend/src/main/java/com/example/backend/repository/Product/ProductRepository.java
@@ -11,6 +11,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -48,7 +49,7 @@ public interface ProductRepository extends JpaRepository<Product, Long>, AdminPr
     @Modifying
     @Transactional
     @Query("UPDATE Product p SET p.previousPrice = :previousContractPrice WHERE p.productId = :recentlyProductId")
-    void updatePreviousPrice(@Param("recentlyProductId") Long recentlyProductId, @Param("previousContractPrice") Long previousContractPrice);
+    void updatePreviousPrice(@Param("recentlyProductId") Long recentlyProductId, @Param("previousContractPrice") BigDecimal previousContractPrice);
 
     // 해당 Id의 변동률 계산
     @Modifying
@@ -60,7 +61,7 @@ public interface ProductRepository extends JpaRepository<Product, Long>, AdminPr
     @Modifying
     @Transactional
     @Query("UPDATE Product p SET p.latestPrice = :latestPrice WHERE p.productId = :productId")
-    void updateLatestPrice(@Param("productId") Long productId, @Param("latestPrice") Long latestPrice);
+    void updateLatestPrice(@Param("productId") Long productId, @Param("latestPrice") BigDecimal latestPrice);
 
     @Modifying
     @Transactional
